+++ src/util/index.ts
@@ -2,3 +2,4 @@ export * from './create-logger';
 export * from './get-headers';
 export * from './get-header-config';
 export * from './extract-field-and-value';
+export * from './fill-optional-fields';
